31.1.2025
can see exercises and sets
have to choose category and finish workout
on finish add a point and clear the table. Maybe add a visible column to the table so when finished with the workout the exercises will be set to visible=false. on finish have to check as comnpleted the same day for the calendar